+++ NEWS.md
@@ -4,6 +4,11 @@ output: html_document
 ---
 
 
+
+
+* New 4-letter abbreviations for non-standard ISO country codes.
+
+
 # IEATools 0.1.63 (2022-04-01) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6407190.svg)](https://doi.org/10.5281/zenodo.6407190)
 
 * Added `non_energy_flows` to `fd_sectors`